requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic
python-multipart
genai
